# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * RFC 7807. O texto que o usuário vê é escolhido pelo cliente a partir do `code`; o servidor não manda texto de produto.
 */
export interface Problem {
  type: string;
  title: string;
  status: number;
  /** Identificador estável que o cliente compara */
  code: string;
  detail?: string;
}

/**
 * Valores derivados de lib/core — a fonte única da regra de domínio
 */
export type WorkspaceStatus = typeof WorkspaceStatus[keyof typeof WorkspaceStatus];


export const WorkspaceStatus = {
  sem_edital: 'sem_edital',
  aguardando_upload: 'aguardando_upload',
  extraindo_edital: 'extraindo_edital',
  aguardando_revisao_edital: 'aguardando_revisao_edital',
  diagnostico_pendente: 'diagnostico_pendente',
  diagnostico_em_andamento: 'diagnostico_em_andamento',
  plano_quinzenal_pendente: 'plano_quinzenal_pendente',
  estudando: 'estudando',
  erro: 'erro',
} as const;

/**
 * Mapeado do status; não é campo persistido
 */
export type ImportStatus = typeof ImportStatus[keyof typeof ImportStatus];


export const ImportStatus = {
  pending: 'pending',
  parsing: 'parsing',
  completed: 'completed',
  error: 'error',
} as const;

/**
 * De onde veio o edital. `none` = ainda não importado
 */
export type SourceMode = typeof SourceMode[keyof typeof SourceMode];


export const SourceMode = {
  file: 'file',
  text: 'text',
  none: 'none',
} as const;

export interface DayAvailability {
  /**
     * @minimum 0
     * @maximum 6
     */
  weekday: number;
  /** @minimum 0 */
  minutes: number;
}

export interface WeeklyAvailability {
  days: DayAvailability[];
  /** @minimum 0 */
  maxSessionMinutes: number;
}

/**
 * @minLength 1
 * @maxLength 64
 * @pattern ^[A-Za-z0-9_-]+$
 */
export type CargoId = string;

export interface Cargo {
  id: CargoId;
  name: string;
  examDate?: string | null;
  period?: string | null;
  position: number;
}

export interface CargoCreate {
  /**
     * Opcional. Aceito SÓ na criação; depois disso o id é imutável, e `CargoPatch` não o inclui. Sem ele o servidor gera um.
     * O id do cargo é identidade de DOMÍNIO compartilhada, não detalhe privado do banco: `syllabus_item_cargo` guarda `cargo_id`, e enquanto o módulo de syllabus for local as ligações que ele grava precisam casar com o cargo que a API conhece.
     * É único dentro do workspace, não globalmente — dois concursos têm cada um o seu `c1`. Duplicar dentro do mesmo workspace responde 409.
     */
  id?: CargoId;
  /** @minLength 1 */
  name: string;
  examDate?: string | null;
  period?: string | null;
}

export interface CargoPatch {
  /** @minLength 1 */
  name?: string;
  examDate?: string | null;
  period?: string | null;
  position?: number;
  /** Marcar true desmarca os outros cargos do mesmo workspace */
  isSelected?: boolean;
}

export interface Workspace {
  slug: string;
  title: string;
  institution: string;
  /** Texto livre ("Concurso Público"), não enumeração */
  type: string;
  examDate?: string | null;
  cargos: Cargo[];
  /** Derivado de cargo.is_selected. A seleção é propriedade do cargo, não do workspace — guardá-la aqui criaria dependência circular. */
  readonly selectedCargoId?: string | null;
  availability: WeeklyAvailability;
  status: WorkspaceStatus;
  sourceMode: SourceMode;
  sourceFileName?: string | null;
  active: boolean;
  /** Alvo do If-Match. Por workspace, não por documento. */
  readonly version: number;
  /** Função pura do status, computada por nextActionFor. Nunca persistido — um campo gravado seria segunda representação de um fato que o status já carrega. */
  readonly nextAction: string;
  readonly importStatus: ImportStatus;
}

/**
 * Não aceita status nem nenhum derivado. O status inicial é decidido por lib/core a partir de haver edital ou não.
 */
export interface WorkspaceCreate {
  /** Opcional; sem ele o servidor deriva do título por slugify */
  slug?: string;
  /** @minLength 1 */
  title: string;
  institution?: string;
  type?: string;
  examDate?: string | null;
  availability?: WeeklyAvailability;
  sourceMode?: SourceMode;
  sourceFileName?: string | null;
  cargos?: CargoCreate[];
}

/**
 * Subconjunto de campos editáveis. Derivados são recusados com 400, não ignorados — ignorar em silêncio faz o cliente acreditar que escreveu.
 */
export interface WorkspacePatch {
  /** @minLength 1 */
  title?: string;
  institution?: string;
  type?: string;
  examDate?: string | null;
  availability?: WeeklyAvailability;
  status?: WorkspaceStatus;
  sourceMode?: SourceMode;
  sourceFileName?: string | null;
  active?: boolean;
}

/**
 * Sem sessão válida
 */
export type NaoAutenticadoResponse = Problem;

/**
 * Não existe, ou não é do usuário da sessão. As duas situações respondem igual, de propósito.
 */
export type NaoEncontradoResponse = Problem;

/**
 * Forma inválida, ou campo derivado enviado pelo cliente
 */
export type RequisicaoInvalidaResponse = Problem;

/**
 * Falta o cabeçalho If-Match
 */
export type PrecondicaoObrigatoriaResponse = Problem;

/**
 * O If-Match enviado não é a version atual
 */
export type PrecondicaoFalhouResponse = Problem;

/**
 * Falha não tratada. O corpo é fixo: nunca traz mensagem de erro, SQL, stack nem variável de ambiente.
 */
export type ErroInternoResponse = Problem;

/**
 * A `version` que o cliente leu. Ausente responde 428; diferente da atual responde 412. Não existe "esqueci a versão, sobrescreve assim mesmo".
 */
export type IfMatchParameter = string;