# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces/{slug}/cargos:
    parameters:
      - $ref: "#/components/parameters/Slug"
    post:
      operationId: createCargo
      tags: [cargos]
      summary: Acrescenta um cargo ao workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CargoCreate"
      responses:
        "201":
          description: Criado; devolve o workspace inteiro, com a nova version
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "400": { $ref: "#/components/responses/RequisicaoInvalida" }
        "401": { $ref: "#/components/responses/NaoAutenticado" }
        "404": { $ref: "#/components/responses/NaoEncontrado" }
        "500": { $ref: "#/components/responses/ErroInterno" }
  /workspaces/{slug}/cargos/{cargoId}:
    parameters:
      - $ref: "#/components/parameters/Slug"
      - $ref: "#/components/parameters/CargoId"
    patch:
      operationId: patchCargo
      tags: [cargos]
      summary: Altera um cargo
      description: >
        Marcar `isSelected: true` desmarca os outros na mesma transação — o índice
        único parcial do banco recusa dois selecionados no mesmo workspace.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CargoPatch"
      responses:
        "200":
          description: Alterado; devolve o workspace inteiro
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "400": { $ref: "#/components/responses/RequisicaoInvalida" }
        "401": { $ref: "#/components/responses/NaoAutenticado" }
        "404": { $ref: "#/components/responses/NaoEncontrado" }
        "500": { $ref: "#/components/responses/ErroInterno" }
    delete:
      operationId: deleteCargo
      tags: [cargos]
      summary: Remove um cargo
      description: >
        Remover o cargo selecionado promove o de menor `position` — a interface
        sempre mostra um cargo. Remover o último é recusado com 409.
      responses:
        "200":
          description: Removido; devolve o workspace inteiro
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "401": { $ref: "#/components/responses/NaoAutenticado" }
        "404": { $ref: "#/components/responses/NaoEncontrado" }
        "409":
          description: É o último cargo do workspace
          content:
            application/problem+json:
              schema: { $ref: "#/components/schemas/Problem" }
        "500": { $ref: "#/components/responses/ErroInterno" }
components:
  parameters:
    Slug:
      name: slug
      in: path
      required: true
      description: Slug do workspace, único por usuário
      schema: { type: string }
    CargoId:
      name: cargoId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/CargoId" }
...
  responses:
    NaoAutenticado:
      description: Sem sessão válida
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    NaoEncontrado:
      description: >
        Não existe, ou não é do usuário da sessão. As duas situações respondem
        igual, de propósito.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
    RequisicaoInvalida:
      description: Forma inválida, ou campo derivado enviado pelo cliente
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    ErroInterno:
      description: >
        Falha não tratada. O corpo é fixo: nunca traz mensagem de erro, SQL, stack
        nem variável de ambiente.
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    Problem:
      type: object
      description: >
        RFC 7807. O texto que o usuário vê é escolhido pelo cliente a partir do
        `code`; o servidor não manda texto de produto.
      required: [type, title, status, code]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        code:
          type: string
          description: Identificador estável que o cliente compara
        detail: { type: string }
    WorkspaceStatus:
      type: string
      description: Valores derivados de lib/core — a fonte única da regra de domínio
      enum:
        - sem_edital
        - aguardando_upload
        - extraindo_edital
        - aguardando_revisao_edital
        - diagnostico_pendente
        - diagnostico_em_andamento
        - plano_quinzenal_pendente
        - estudando
        - erro
    ImportStatus:
      type: string
      description: Mapeado do status; não é campo persistido
      enum: [pending, parsing, completed, error]
    SourceMode:
      type: string
      description: De onde veio o edital. `none` = ainda não importado
      enum: [file, text, none]
    DayAvailability:
      type: object
      required: [weekday, minutes]
      properties:
        weekday:
          type: integer
          minimum: 0
          maximum: 6
        minutes:
          type: integer
          minimum: 0
    WeeklyAvailability:
      type: object
      required: [days, maxSessionMinutes]
      properties:
        days:
          type: array
          items: { $ref: "#/components/schemas/DayAvailability" }
        maxSessionMinutes:
          type: integer
          minimum: 0
    CargoId:
      type: string
      minLength: 1
      maxLength: 64
      # Rótulo curto do domínio (`c1`, `c2`), não UUID. Restrito a letras, números,
      # hífen e sublinhado: o id viaja em URL de rota e em chave de armazenamento,
      # e aceitar qualquer texto abriria caminho para barra, ponto e espaço.
      pattern: "^[A-Za-z0-9_-]+$"
    Cargo:
      type: object
      required: [id, name, position]
      properties:
        id: { $ref: "#/components/schemas/CargoId" }
        name: { type: string }
        examDate: { type: string, format: date, nullable: true }
        period: { type: string, nullable: true }
        position: { type: integer }
    CargoCreate:
      type: object
      required: [name]
      properties:
        id:
          allOf: [{ $ref: "#/components/schemas/CargoId" }]
          description: >
            Opcional. Aceito SÓ na criação; depois disso o id é imutável, e
            `CargoPatch` não o inclui. Sem ele o servidor gera um.

            O id do cargo é identidade de DOMÍNIO compartilhada, não detalhe
            privado do banco: `syllabus_item_cargo` guarda `cargo_id`, e enquanto
            o módulo de syllabus for local as ligações que ele grava precisam
            casar com o cargo que a API conhece.

            É único dentro do workspace, não globalmente — dois concursos têm cada
            um o seu `c1`. Duplicar dentro do mesmo workspace responde 409.
        name: { type: string, minLength: 1 }
        examDate: { type: string, format: date, nullable: true }
        period: { type: string, nullable: true }
    CargoPatch:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        examDate: { type: string, format: date, nullable: true }
        period: { type: string, nullable: true }
        position: { type: integer }
        isSelected:
          type: boolean
          description: Marcar true desmarca os outros cargos do mesmo workspace
    Workspace:
      type: object
      required:
        - slug
        - title
        - institution
        - type
        - cargos
        - availability
        - status
        - sourceMode
        - active
        - version
        - nextAction
        - importStatus
      properties:
        slug: { type: string }
        title: { type: string }
        institution: { type: string }
        type:
          type: string
          description: Texto livre ("Concurso Público"), não enumeração
        examDate: { type: string, format: date, nullable: true }
        cargos:
          type: array
          items: { $ref: "#/components/schemas/Cargo" }
        selectedCargoId:
          type: string
          nullable: true
          readOnly: true
          description: >
            Derivado de cargo.is_selected. A seleção é propriedade do cargo, não do
            workspace — guardá-la aqui criaria dependência circular.
        availability: { $ref: "#/components/schemas/WeeklyAvailability" }
        status: { $ref: "#/components/schemas/WorkspaceStatus" }
        sourceMode: { $ref: "#/components/schemas/SourceMode" }
        sourceFileName: { type: string, nullable: true }
        active: { type: boolean }
        version:
          type: integer
          readOnly: true
          description: Alvo do If-Match. Por workspace, não por documento.
        nextAction:
          type: string
          readOnly: true
          description: >
            Função pura do status, computada por nextActionFor. Nunca persistido —
            um campo gravado seria segunda representação de um fato que o status já
            carrega.
        importStatus:
          $ref: "#/components/schemas/ImportStatus"
          readOnly: true
...