# lib/api-zod/src/generated/types/cargo.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CargoId } from './cargoId';

export interface Cargo {
  id: CargoId;
  name: string;
  examDate?: Date | null;
  period?: string | null;
  position: number;
}